backend-moneylover/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8

djangorestframework>=3.10
djangorestframework-simplejwt>=4
django-filter==2.4





